Exclude pom.xml files from triggering pipelines
As a workaround to prevent every pipeline in the repo from
triggering when we increment the verison of azure-core we are
excluding pom.xml file only changes from triggering the other
pipelines.

The expectation is that most real changes, outside of our version
increments, actually change other files in a service directory as
well so they will still trigger. If there is ever a case where only
a pom.xml file is changed then we will have to manually trigger
the pipeline via a comment to test it.